backend/model_api/requirements.txt
flask==2.0.1
werkzeug==2.0.1
numpy==1.23.5
pandas==1.5.3
tensorflow==2.12.0
pillow==9.5.0
flask-cors==3.0.10
h5py==3.13.0
